+++ resources/views/holidays/list.blade.php
@@ -143,7 +143,7 @@
                                 <div class="item-meeting">
                                     <p>{{$holiday->date}}</p>
                                 </div>
-                                <a href="delete_holidays/{{$holiday->id_holiday}}"><button type="button" class="btn btn-danger">Deletar</button></a>
+                                <a href="delete_holidays/{{$holiday->id_holiday}}" onclick="return confirm('Deseja deletar?')"><button type="button" class="btn btn-danger">Deletar</button></a>
                                 <a href="view_holidays/{{$holiday->id_holiday}}"><button type="button" class="btn btn-edit">Editar</button></a>
                             </div>
                             <!--Item-->
