+++ resources/views/officies/list.blade.php
@@ -143,7 +143,7 @@
                                 <div class="item-meeting">
                                     <p>{{$office->description}}</p>
                                 </div>
-                                <a href="delete_officies/{{$office->id_office}}"><button type="button" class="btn btn-danger">Deletar</button></a>
+                                <a href="delete_officies/{{$office->id_office}}" onclick="return confirm('Deseja deletar?')"><button type="button" class="btn btn-danger">Deletar</button></a>
                                 <a href="view_officies/{{$office->id_office}}"><button type="button" class="btn btn-edit">Editar</button></a>
                             </div>
                             <!--Item-->
